Remove unnecessary components
