+++ bmh_lims/frontend/src/components/Buttons/Styles.js
@@ -7,6 +7,7 @@ const ButtonBase = css`
     padding: 0.5em 1em;
     border-radius: 5px;
     min-height: 1.5em;
+    width: ${props => props.width || 'unset'};
 `
 
 const FilledButtonBase = styled.button`
