--- bmh_lims/frontend/src/components/Checkbox/Styles.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components'
-import { RiCheckboxBlankCircleLine, RiCheckboxCircleFill} from 'react-icons/ri'
-
-export const CheckboxContainer = styled.div`
-    display: inline-block;
-    width: 100%;
-    text-align: center;
-`
