+++ bmh_lims/frontend/src/components/Checkbox/index.js
@@ -1,10 +1,9 @@
 import React from 'react'
-import { CheckboxContainer } from './Styles'
 import { RiCheckboxBlankCircleLine, RiCheckboxCircleFill} from 'react-icons/ri'
 
 export const Checkbox = ({checked, onChangeHandler, colour}) => {
     return (
-        <CheckboxContainer onClick={(e) => {onChangeHandler(e)}}>
+        <div onClick={(e) => {onChangeHandler(e)}}>
             {checked ? 
             <RiCheckboxCircleFill style={{
                 fill: colour,
@@ -14,6 +13,6 @@ export const Checkbox = ({checked, onChangeHandler, colour}) => {
                 fill: colour,
                 verticalAlign: 'middle'
             }}/>}
-        </CheckboxContainer>
+        </div>
     )
 }
