+++ bmh_lims/frontend/src/components/Table/Styles.js
@@ -3,6 +3,14 @@ import styled from 'styled-components'
 // table: sharp corners with shadows
 // headers are divs (or add scroll to only t-body)
 
+export const TableOuterContainer = styled.div`
+    width: ${props => props.width || '100%'};
+    min-height: ${props => props.minHeight || '20%'};
+    height: ${props => props.height || '50%'};
+    position: relative;
+    margin: ${props => props.margin || '0 0 1% 0'};
+`
+
 export const TableContainer = styled.div`
     overflow-x: auto;
     overflow-y: auto;
