+++ bmh_lims/frontend/src/components/Table/index.js
@@ -8,7 +8,8 @@ import {
     HeaderSeparator,
     Table,
     Row,
-    TableContainer
+    TableContainer,
+    TableOuterContainer
 } from './Styles'
 
 import { CheckboxColumn, MappedCell } from './components/'
@@ -36,49 +37,51 @@ const StickyTable = ({theme, headers, content, valueUpdateHandler, isSelectable,
     const sanitizedValue = value => isDate(value) ? formatDate(new Date(value)) : value
 
     return (
-        <TableContainer>
-            {isSelectable && 
-            <CheckboxColumn
-            onSelect={onSelect}
-            numRows={content.length}
-            colour={theme.colour4}
-            />}
-            <Table>
-                <HeaderSeparator>
-                    <Row>
-                        {headers.map((header, hidx) => (
-                        <HeaderCell key={`header-${hidx}`}>
-                            <MappedCell 
-                            header={true}
-                            key={`content-${header}`}
-                            updateColWidths={(width) => {
-                                colWidths[hidx] = width
-                                updateColWidths([...colWidths])
-                            }}>
-                                {`${header.split('_').join(' ')}`}
-                            </MappedCell>
-                        </HeaderCell>))}
-                    </Row>
-                </HeaderSeparator>
-                <BodySeparator>
-                    {content.map((row, ridx) => (
-                    <Row key={`row-${ridx}`}>{
-                        row.map((item, idx) => (
-                            <BodyCell key={`cell-${ridx}-${idx}`}>
-                                <Content width={colWidths[idx]}>
-                                    <BodyContent
-                                    type='text'
-                                    value={sanitizedValue(item) || ''}
-                                    onChange={valueUpdateHandler ? valueUpdateHandler(idx, ridx): (e) => {}}
-                                    readOnly={!isEditable}
-                                    />
-                                </Content>
-                            </BodyCell>)
-                        )}
-                    </Row>))}
-                </BodySeparator>
-            </Table>
-        </TableContainer>
+        <TableOuterContainer>
+            <TableContainer>
+                {isSelectable && 
+                <CheckboxColumn
+                onSelect={onSelect}
+                numRows={content.length}
+                colour={theme.colour4}
+                />}
+                <Table>
+                    <HeaderSeparator>
+                        <Row>
+                            {headers.map((header, hidx) => (
+                            <HeaderCell key={`header-${hidx}`}>
+                                <MappedCell 
+                                header={true}
+                                key={`content-${header}`}
+                                updateColWidths={(width) => {
+                                    colWidths[hidx] = width
+                                    updateColWidths([...colWidths])
+                                }}>
+                                    {`${header.split('_').join(' ')}`}
+                                </MappedCell>
+                            </HeaderCell>))}
+                        </Row>
+                    </HeaderSeparator>
+                    <BodySeparator>
+                        {content.map((row, ridx) => (
+                        <Row key={`row-${ridx}`}>{
+                            row.map((item, idx) => (
+                                <BodyCell key={`cell-${ridx}-${idx}`}>
+                                    <Content width={colWidths[idx]}>
+                                        <BodyContent
+                                        type='text'
+                                        value={sanitizedValue(item) || ''}
+                                        onChange={valueUpdateHandler ? valueUpdateHandler(idx, ridx): (e) => {}}
+                                        readOnly={!isEditable}
+                                        />
+                                    </Content>
+                                </BodyCell>)
+                            )}
+                        </Row>))}
+                    </BodySeparator>
+                </Table>
+            </TableContainer>
+        </TableOuterContainer>
 )}
 
 export default StickyTable
