+++ bmh_lims/frontend/src/pages/Workflows/sections/index.js
@@ -1,0 +1,1 @@
+export { AssignSection } from './AssignWorkflows/'
